Some debugging
Tried to fix the Energy Expenditure algorithm. Noticed that I was subtracting 1.3 everytime I sampled and not once. Should be fixed.
Tested with a light jog and found readings to be about 10times higher than expected. Will do research to try and minimize error. May need new (or homemade) model.